@@ -2,7 +2,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<groupId>org.imixs.workflow</groupId>
 	<artifactId>imixs-adapters</artifactId>
-	<version>2.3.3-SNAPSHOT</version>
+	<version>3.0.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
 	<name>Imxis Adapters</name>
 	<modules>
@@ -18,12 +18,12 @@
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<org.imixs.workflow.version>5.2.13</org.imixs.workflow.version>
-		<org.imixs.marty.version>4.2.0</org.imixs.marty.version>
-		<org.imixs.archive.version>2.2.11</org.imixs.archive.version>
-		<jakarta.version>8.0</jakarta.version>
+		<org.imixs.workflow.version>6.0.3-SNAPSHOT</org.imixs.workflow.version>
+		<org.imixs.marty.version>5.0.0-SNAPSHOT</org.imixs.marty.version>
+		<org.imixs.archive.version>3.0.0-SNAPSHOT</org.imixs.archive.version>
+		<jakarta.version>10.0.0</jakarta.version>
 		<microprofile.version>3.0</microprofile.version>
-		<apache.pdfbox.version>2.0.19</apache.pdfbox.version>
+		<apache.pdfbox.version>2.0.21</apache.pdfbox.version>
 		<org.imixs.jwt.version>1.1.0</org.imixs.jwt.version>
 	</properties>
 
@@ -83,8 +83,8 @@
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.8.1</version>
 				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
+					<source>11</source>
+					<target>11</target>
 				</configuration>
 			</plugin>
 
@@ -284,18 +284,24 @@
 	<dependencies>
 		<!-- Jakarta EE -->
 		<dependency>
-			<groupId>javax</groupId>
-			<artifactId>javaee-api</artifactId>
-			<version>${jakarta.version}</version>
-			<scope>provided</scope>
+		    <groupId>jakarta.platform</groupId>
+		    <artifactId>jakarta.jakartaee-api</artifactId>
+		    <version>${jakarta.version}</version>
+		    <scope>provided</scope>
 		</dependency>
 		<!-- JDK 11 - XML Bind -->
 		<dependency>
-		    <groupId>jakarta.xml.bind</groupId>
-		    <artifactId>jakarta.xml.bind-api</artifactId>
-		    <version>2.3.3</version>
-		    <scope>provided</scope>
+		    <groupId>org.glassfish.jaxb</groupId>
+		    <artifactId>jaxb-runtime</artifactId>
+		    <version>3.0.0</version>
+		    <scope>test</scope>
 		</dependency>
+    	<dependency>
+	       <groupId>jakarta.xml.bind</groupId>
+	       <artifactId>jakarta.xml.bind-api</artifactId>
+	       <version>3.0.0</version>
+	       <scope>test</scope>
+	    </dependency>	
 		
 		<!-- Microprofile -->
 		<dependency>
