@@ -34,8 +34,8 @@ import java.util.Map;
 import java.util.logging.Logger;
 import java.util.regex.Pattern;
 
-import javax.ejb.EJB;
-import javax.xml.bind.JAXBException;
+import jakarta.ejb.EJB;
+import jakarta.xml.bind.JAXBException;
 import javax.xml.transform.TransformerException;
 
 import org.imixs.workflow.FileData;
